requirements.txt
Django>=5.1,<6.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
